+++ src/main/java/io/github/ryuu/adventurecraft/rendering/RenderArrowBomb.java
@@ -1,7 +1,10 @@
 package io.github.ryuu.adventurecraft.rendering;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.render.Tessellator;
 import net.minecraft.client.render.entity.ArrowRenderer;
+import net.minecraft.entity.projectile.Arrow;
 import net.minecraft.util.maths.MathsHelper;
 import org.lwjgl.opengl.GL11;
 
@@ -11,9 +14,9 @@ public class RenderArrowBomb extends ArrowRenderer {
     public void render(MixinArrow entity, double x, double y, double z, float f, float f1) {
         this.bindTexture("/item/arrows.png");
         GL11.glPushMatrix();
-        GL11.glTranslatef((float) x, (float) y, (float) z);
-        GL11.glRotatef((float) (entity.prevYaw + (entity.yaw - entity.prevYaw) * f1 - 90.0f), 0.0f, 1.0f, 0.0f);
-        GL11.glRotatef((float) (entity.prevPitch + (entity.pitch - entity.prevPitch) * f1), 0.0f, 0.0f, 1.0f);
+        GL11.glTranslatef((float) ((float) x), (float) ((float) y), (float) ((float) z));
+        GL11.glRotatef((float) (entity.prevYaw + (entity.yaw - entity.prevYaw) * f1 - 90.0f), (float) 0.0f, (float) 1.0f, (float) 0.0f);
+        GL11.glRotatef((float) (entity.prevPitch + (entity.pitch - entity.prevPitch) * f1), (float) 0.0f, (float) 0.0f, (float) 1.0f);
         Tessellator tessellator = Tessellator.INSTANCE;
         int i = 0;
         float f2 = 0.0f;
@@ -25,23 +28,23 @@ public class RenderArrowBomb extends ArrowRenderer {
         float f8 = (float) (5 + i * 10) / 32.0f;
         float f9 = (float) (10 + i * 10) / 32.0f;
         float f10 = 0.05625f;
-        GL11.glEnable(32826);
+        GL11.glEnable((int) 32826);
         float f11 = (float) entity.shake - f1;
         if (f11 > 0.0f) {
             float f12 = -MathsHelper.sin(f11 * 3.0f) * f11;
-            GL11.glRotatef(f12, 0.0f, 0.0f, 1.0f);
+            GL11.glRotatef((float) f12, (float) 0.0f, (float) 0.0f, (float) 1.0f);
         }
-        GL11.glRotatef(45.0f, 1.0f, 0.0f, 0.0f);
-        GL11.glScalef(f10, f10, f10);
-        GL11.glTranslatef(-4.0f, 0.0f, 0.0f);
-        GL11.glNormal3f(f10, 0.0f, 0.0f);
+        GL11.glRotatef((float) 45.0f, (float) 1.0f, (float) 0.0f, (float) 0.0f);
+        GL11.glScalef((float) f10, (float) f10, (float) f10);
+        GL11.glTranslatef((float) -4.0f, (float) 0.0f, (float) 0.0f);
+        GL11.glNormal3f((float) f10, (float) 0.0f, (float) 0.0f);
         tessellator.start();
         tessellator.vertex(-7.0, -2.0, -2.0, f6, f8);
         tessellator.vertex(-7.0, -2.0, 2.0, f7, f8);
         tessellator.vertex(-7.0, 2.0, 2.0, f7, f9);
         tessellator.vertex(-7.0, 2.0, -2.0, f6, f9);
         tessellator.draw();
-        GL11.glNormal3f(-f10, 0.0f, 0.0f);
+        GL11.glNormal3f((float) (-f10), (float) 0.0f, (float) 0.0f);
         tessellator.start();
         tessellator.vertex(-7.0, 2.0, -2.0, f6, f8);
         tessellator.vertex(-7.0, 2.0, 2.0, f7, f8);
@@ -49,8 +52,8 @@ public class RenderArrowBomb extends ArrowRenderer {
         tessellator.vertex(-7.0, -2.0, -2.0, f6, f9);
         tessellator.draw();
         for (int j = 0; j < 4; ++j) {
-            GL11.glRotatef(90.0f, 1.0f, 0.0f, 0.0f);
-            GL11.glNormal3f(0.0f, 0.0f, f10);
+            GL11.glRotatef((float) 90.0f, (float) 1.0f, (float) 0.0f, (float) 0.0f);
+            GL11.glNormal3f((float) 0.0f, (float) 0.0f, (float) f10);
             tessellator.start();
             tessellator.vertex(-8.0, -2.0, 0.0, f2, f4);
             tessellator.vertex(8.0, -2.0, 0.0, f3, f4);
@@ -58,7 +61,7 @@ public class RenderArrowBomb extends ArrowRenderer {
             tessellator.vertex(-8.0, 2.0, 0.0, f2, f5);
             tessellator.draw();
         }
-        GL11.glDisable(32826);
+        GL11.glDisable((int) 32826);
         GL11.glPopMatrix();
     }
 }
