+++ src/main/java/io/github/ryuu/adventurecraft/util/Selection.java
@@ -1,5 +1,8 @@
 package io.github.ryuu.adventurecraft.util;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
 class Selection {
 
     int oneX;
@@ -33,11 +36,11 @@ class Selection {
         ItemCursor.twoX = this.twoX;
         ItemCursor.twoY = this.twoY;
         ItemCursor.twoZ = this.twoZ;
-        ItemCursor.minX = Math.min(this.oneX, this.twoX);
-        ItemCursor.minY = Math.min(this.oneY, this.twoY);
-        ItemCursor.minZ = Math.min(this.oneZ, this.twoZ);
-        ItemCursor.maxX = Math.max(this.oneX, this.twoX);
-        ItemCursor.maxY = Math.max(this.oneY, this.twoY);
-        ItemCursor.maxZ = Math.max(this.oneZ, this.twoZ);
+        ItemCursor.minX = Math.min((int) this.oneX, (int) this.twoX);
+        ItemCursor.minY = Math.min((int) this.oneY, (int) this.twoY);
+        ItemCursor.minZ = Math.min((int) this.oneZ, (int) this.twoZ);
+        ItemCursor.maxX = Math.max((int) this.oneX, (int) this.twoX);
+        ItemCursor.maxY = Math.max((int) this.oneY, (int) this.twoY);
+        ItemCursor.maxZ = Math.max((int) this.oneZ, (int) this.twoZ);
     }
 }
